Search Patrons - Make patron search more forgiving.
Patron search has only been searching fields for names
that either match or begin with the given search terms.

This is too strict a search.
For example, take the name "Billie Jo Robinson", "Billie Jo"
being the firstname, and "Robinson" being the last name. If one
were to search for "Billie Jo", this patron would not appear in
the search results.

 To remedy this, the search
type has been changed from 'begins_with' to 'contain'.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>