Circulation - Disable Barcode Field If Item Needs Confirmation
When issuing items, if a confirmation screen comes up,
another item can still be scanned into the barcode field.

If a librarian does not notice the confirmation dialog,
he or she may continue scanning items without confirming
the issue and thus give a patron an item without it actually
being issued to the patron.

This patch disables the barcode field until the confirmation
question is answered.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>