Fix for Bug 4271, Offer choice of destination for save biblio operation
A follow-up to the fix for Bug 4121, the user should be able to control whether
the action of saving a bibliographic record takes them to the add item screen
or the detail screen, independently of whether they are adding or modifying the
record.

The addition of a "split button" would leave the default action in place if the
user clicked the "main" part of the button. Clicking on the menu would give the
user specific choices.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>