Fixed Patron Search Results
By default, the patron search was pulling results in *Descending* order,
then reordering the results on the page using javascript in *Ascending* order.

This was causing screwy results where the entire list of results of multiple
pages was not in complete alphabetical order.

Fixed by changing the SQL to pull in Ascending order.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>