Fix for Bug 4416, renew all and return all buttons too close together
This patch removes the "return all" button altogether. Returning all items
is a step which one wouldn't take lightly or even regularly. Since the
interface has a means of selecting all checkboxes in the "Check in" column,
removing the "Return all" button wouldn't eliminate the means of doing so, it
would just require more deliberation.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>