Bug 4470 Patron search result pagination bar
Obviously, Koha global pagination function could be improved. This patch
do something which belongs to this function. A CPAN module like
Data::Page could help or be an inspiration... 3.4?

There is also room for improvement in they way resultset are
constructed. The whole set is retrieved and just a subset is displayed.
It impacts performances for libraries with a great number of patrons.
The right solution would be:

  - to find the resultset size with SELECT COUNT(*)
  - to retrieved the exact subset with LIMIT and OFFSET SQL arguments

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>