add EUC-KR to list of support Z39.50 character encodings
EUC-KR (an older character encoding for Korean) is used by the
Z39.50 server of the National Assembly Library of the Republic
of Korea.

Note that this requires that a suitable version of iconv,
one that supports mapping from EUC-KR to UTF-8,
be present on the Koha server.

Thanks to a person from the Korea Society for reporting the
issue on #koha.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>