Fix for Bug 4849 - Cannot delete records from Labeled MARC view
The template checks for a "count" variable (item count) to see
whether it should allow deletion of the biblio record. For some
reason labeledMARCdetail.pl hard-coded this as 1, which meant
the template always thought was 1 item.

This patch also adds an item count check on the display of
"Delete all items," disabling it if there are no items.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>