Bug 4867 - cannot clone subfields in marc editor, other clone/unclone weirdness
When the clone/unclone plus/minus were replaced with images the <span> tags were inadvertantly deleted
causing the js to loose its ability to identify the clone/unclone links and properly udpate the indexes.
The resulting incorrect indexes caused various weirdnesses to result when one attempted to clone/unclone
subfields.

This patch simply adds classes to the two links and updates the js to identify the links based on class.

Thanks to owen for his help in chasing this down.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Galen Charlton <gmcharlt@gmail.com>