Fix for Bug 4218 - Staff client detail page does not show item hold status
Showing "on hold" in the holdings list (as we used to do) is misleading
if it is a bibilo-level hold. However seeing that a title has holds
is important information for this view. This patch adds a line to the
biblio-level information display with a count of the title's outstanding
holds.

Signed-off-by: Nicole Engard <nengard@bywatersolutions.com>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>