Bug 5203 - Creating a label template causes fatal error if description field is empty
This is due to the underlying script not having any defaults hard coded which results
in skewed hash elements when the parameter array is coerced into a hash at a later point.

This also occurs in layouts and profiles. This patch adds defaults to all three.

No documentation changes are necessary.