Fix Bug 5261 broken link in Local Use sysprefs
This patch eliminates the sysprefs-menu.inc include file and changes the
systempreferences.pl and systempreferences.tmpl files to work with the
prefs-menu.inc instead. This will centralize the syspref tabs and make it
easier to modify tabs in the future if necessary.

This commit also changes the default tab to Acquisitions, since the Local Use
tab does not work with preferences.pl at present.

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Galen Charlton <gmcharlt@gmail.com>