Fix for Bug 5051, Renewal due date doesn't always show on patron Checkout tab
Display of the renewal date footer in the template was conditional upon
the existence of checkouts from today.

In order to fix this bug it was necessary to repeat the markup for the
table footer twice in the template, so I created an include file for it.
The same include file can now be used in moremember.tmpl.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>