Adding id tags to some OPAC templates.
Many of the templates in OPAC, which deal with users data eg opac-account.tmpl, opac-passwd.tmpl etc do not have a id for the container div,
like opac-user.tmpl has an id userdetail. Having these id's makes it easier to customize with css.
This patch adds id's to most of the <div class="container"> tags.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>