Fix for Bug 3659, Add place hold option from patron checkout tab
Adds a button on patron-related pages, "Search to hold"
Clicking search to hold sets a cookie with the patron's
borrowernumber and sends the user to the search page.

On subsequent search results pages the user will have
the option to place holds specifically for the remembered
patron. This works on the search results page (single
and multiple hold) and on the detail page.

The saved cookie will time out after 10 minutes or
be erased when a new patron is loaded by circulation.pl

New jQuery plugin added: jquery.cookie.min.js

Revision fixes some markup and corrects an error that would lead
to the hold being initiated for the remembered patron when this
was not intended.

Signed-off-by: Liz Rea <lrea@nekls.org>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>