Bug 5931 Paging and sorting saved reports table
On Reports > Saved reports page, a table shows all saved reports.
Table's rows are sorted by report IDs. When there are a lot reports,
this table is difficult to read.

This enhancement add to Saved report page a paging bar and sort options on
column names, like in the framework table.

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>
Signed-off-by: Liz Rea <lrea@nekls.org>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>