Bug 5917 follow-up: Fix errors on OPAC results
This patch fixes the following problems with the OPAC search results:
* The facet links were constructed wrong following the change to
  Template::Toolkit

This patch also changes the OPAC sorting dropdown to work in the same way as the
staff dropdown. This seemed preferable to having the two use different variables
for deciding which option to select, and it is unclear why Search.pm was not
passing the appropriate variables to the staff client.

Signed-off-by: Jared Camins-Esakov <jcamins@bywatersolutions.com>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>