Bug 6380 - Z39.50 Cataloging Search needs a way to try search again, even if results are found
This patch adds a nice button to the results page to "Try Another Search" even if results are found. It also makes the "no results found" try again link a proper button.

To test:
Do a search for something that will give results -- there should now be a button to "Try Another Search"

Do a search for something that returns no results -- The link for "Try Another Search" should now be a proper button.

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
[EDIT: Form slightly changed in order to remember biblio and framework. Link is no official html method value btw.]

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>