Fix for Bug 6402, Lists sorted by year appear to be empty
If you create a list, populate it with titles, and then change the sort to
"year" it will appear as if your list is empty. The SQL in VirtualShelves.pm's
GetShelfContents tries to accept "year" as a sort by but 'year' is not a valid
column name.

The SQL pulls biblioitems.publicationyear. Changing this to
'biblioitems.publicationyear as year,'

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>