Bug 6316 - Initial commit for es-ES translated scripts
After some time having doubts on how to start a spanish translation
of SQL files, I decided to submit a copy of 'en' directory, called
'es-ES' so people can start translating.
I hope this is Ok. if not, hopefully this will lead us to some advice
on how to do it :-D

We need spanish-speaking people to help on translating!
Regards
To+