Fix for Bug 6586 - calendar showing the wrong event type when editing
This bug must have been present for a while. There wasn't proper
handling of repeating yearly holidays when it came to the label.

To be clear, I believe this bug only affected holidays which repeat
yearly on the same date.

Instead of correcting the label to match the existing scheme I
added a new color code and label so that there are now two styles
for yearly and weekly repeating holdiays. This makes it more
clear what kind of repeating holiday you're looking at (rather
than having one color code for both).

After applying this patch, new and existing holidays which
repeat yearly on the same day should be color-coded orange and have
a 'Holiday repeating yearly' label. Weekly holidays are now
labeled 'Holiday repeating weekly.'

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>