Fix for Bug 6695 - Layout of patron category add/edit form slightly broken
The markup for the fields for enrollment period "in months" and "until date"
was non-standard I think because the two fields are supposed to be visually
and functionally linked (having the same purpose and being mutually
exclusive).

This patch changes the markup so that the two fields share a fieldset
having "choose one" as the legend. The "hint" is now hidden unless
it is triggered by the javascript validation on submit.

I wanted to trigger the validation of the two fields onblur instead
of just onsubmit, but found that the datepicker widget doesn't
consistently trigger a change or a blur event on that field.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
- Layout problems are fixed
- Layout is more clear about choosing either 'In months' or 'Until date'
- Clear error message shows when you enter values in both fields

Signed-off-by: Ian Walls <ian.walls@bywatersolutions.com>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>