Bug 6690: If multiple issues are received in serials-edit.pl, changing the status of any issue changes the date of the first issue.
If multiple issues are in the list, changing the status of any issue in
the table would always change the "Expected on" date of the first issue.
Modified the javascript so the date is modified on the correct row.

Reimplementation of Patch for 3.4.x
Patch for 3.2.x from Frédérick Capovilla

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>
Signed-off-by: Ian Walls <ian.walls@bywatersolutions.com>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>