6278 Funny translation bug with XSLT subfieldSelect parameters
This patch resolves the situation of translating the parameters at (meaning subfield a and t) of subfieldSelect by replacing at with a_t. This string will not be translated and the xslt function subfieldSelect just ignores the underscore.
It is a pragmatic and easy solution. Theoretically, we should look for a solution in the translation parser. This is much harder and would probably not result in better code in those modules.

Test this patch with the following remarks in mind:
1) This patch deals with XSLT. So you should enable XSLT at OPAC detail.
2) In English or any language not translating the word at: Just check if the links for a MARC 490, 773, 780 or 785 still work at the opac detail page.
3) For a language translating the word at (like Dutch where at becomes bij), you should first rerun your translation process (at least for the OPAC).
4) In that language: Check the results on the vernacular opac detail page for MARC 490, 773, 780 or 785. Without the patch, you would not see a link following e.g. the Series label (with 490 first indicator set to 1). With the patch, you should see a link now.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Pragmatic solution for an annoying problem.

Signed-off-by: Ian Walls <ian.walls@bywatersolutions.com>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>